# desc/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fishing Notes Admin API",
    "version": "0.2.1",
    "description": "管理后台 API 文档，覆盖认证、RBAC、管理员用户、钓点审核、知识专栏、鱼种内容与业务用户管理。\n\n使用说明：\n1. 除登录接口外，其余接口默认都需要 Bearer Token。\n2. 所有成功响应统一为 `{ code, message, data }`，其中 `code = 0` 表示成功。\n3. 列表接口统一支持 `page` 和 `pageSize`，默认 `page=1`、`pageSize=20`。\n"
  },
  "servers": [
    {
      "url": "http://localhost:1002",
      "description": "Local API Server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": [

      ]
    }
  ],
  "paths": {
...
    "/api/v1/rbac/roles": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "获取角色列表",
        "operationId": "listRoles",
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRolesEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "RBAC"
        ],
        "summary": "创建角色",
        "operationId": "createRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/rbac/roles/{roleId}": {
      "put": {
        "tags": [
          "RBAC"
        ],
        "summary": "编辑角色",
        "operationId": "updateRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "RBAC"
        ],
        "summary": "删除角色",
        "operationId": "deleteRole",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/rbac/permissions": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "获取权限列表",
        "operationId": "listPermissions",
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPermissionsEnvelope"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/rbac/users/{userId}/roles": {
      "get": {
        "tags": [
          "RBAC"
        ],
        "summary": "获取指定管理员角色",
        "operationId": "getUserRoles",
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "RBAC"
        ],
        "summary": "更新管理员角色",
        "operationId": "updateUserRoles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRolesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/rbac/roles/{roleId}/permissions": {
      "put": {
        "tags": [
          "RBAC"
        ],
        "summary": "更新角色权限",
        "operationId": "updateRolePermissions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermissionsEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "UserIdPath": {
        "name": "userId",
        "in": "path",
        "required": true,
        "description": "用户 ID",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "RoleIdPath": {
        "name": "roleId",
        "in": "path",
        "required": true,
        "description": "角色 ID",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未认证或认证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "没有权限访问",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "服务端异常",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "RoleItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "type": "integer",
            "format": "int64"
          },
          "permissionCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateRoleRequest": {
        "type": "object",
        "required": [
          "code",
          "name",
          "status",
          "sort",
          "permissionIds"
        ],
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "maxLength": 64
          },
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "status": {
            "type": "integer",
            "format": "int64",
            "enum": [
              0,
              1
            ]
          },
          "sort": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "UpdateRoleRequest": {
        "type": "object",
        "required": [
          "name",
          "status",
          "sort"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64
          },
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "status": {
            "type": "integer",
            "format": "int64",
            "enum": [
              0,
              1
            ]
          },
          "sort": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "PermissionItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "component": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ListRolesResponse": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleItem"
            }
          }
        }
      },
      "ListPermissionsResponse": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionItem"
            }
          }
        }
      },
      "UserRolesResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleItem"
            }
          }
        }
      },
      "UpdateUserRolesRequest": {
        "type": "object",
        "required": [
          "roleIds"
        ],
        "properties": {
          "roleIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "UpdateRolePermissionsRequest": {
        "type": "object",
        "required": [
          "permissionIds"
        ],
        "properties": {
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "RolePermissionsResponse": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int64"
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "SuccessData": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "example": {
          "success": true
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "example": {
          "code": 10000,
          "message": "invalid request"
        }
      },
...
      "ListRolesEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/ListRolesResponse"
          }
        }
      },
      "RoleEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/RoleItem"
          }
        }
      },
      "ListPermissionsEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/ListPermissionsResponse"
          }
        }
      },
      "UserRolesEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/UserRolesResponse"
          }
        }
      },
      "RolePermissionsEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RolePermissionsResponse"
          }
        }
      },
...
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/SuccessData"
          }
        },
        "example": {
          "code": 0,
          "message": "ok",
          "data": {
            "success": true
          }
        }
      }
    }
  }
}